+++ src/const.js
@@ -31,6 +31,7 @@ export const spellEffectMap = {
     'angelic-wings': ['Compendium.pf2e.spell-effects.Item.iZYjxY0qYvg5yPP3'],
     'armor-of-bones': ['Compendium.pf2e.spell-effects.Item.6BjslHgY01cNbKp5'],
     'athletic-rush': ['Compendium.pf2e.spell-effects.Item.57lnrCzGUcNUBP2O'],
+    'true-strike': ['Compendium.pf2e.spell-effects.Item.fpGDAz2v5PG0zUSl'],
 }
 
 export const actionEffectMap = {